Updated remove script, added conditional on nginx files

+++ scripts/remove/lidarr.sh
@@ -2,9 +2,13 @@
 user=$(cut -d: -f1 < /root/.master.info)
 systemctl disable --now lidarr
 
-rm -rf /home/$user/Lidarr
+rm -rf /home/opt/lidarr
 rm -rf /home/$user/.config/Lidarr/
-rm -rf /etc/nginx/apps/lidarr.conf
 rm -rf /install/.lidarr.lock
 rm -rf /etc/systemd/system/lidarr.service
-systemctl reload nginx
+systemctl daemon-reload
+
+if [[ -f /install/.nginx.lock ]]; then
+    rm -rf /etc/nginx/apps/lidarr.conf
+    systemctl reload nginx
+fi
